PycharmProjects/dv_compare_project/gcp_services/functions/compare_data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery[bqstorage,pandas]
pandas
google-cloud-pubsub